% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{addAIQVar}
\alias{addAIQVar}
\title{addAIQVar}
\usage{
addAIQVar(dat)
}
\arguments{
\item{dat}{An existing dataset.}

\item{Vars}{Select variables.}
}
\value{

}
\description{
Add variables from Bounded Structures to existing dataset.
}
