% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{addARTCov}
\alias{addARTCov}
\title{addARTCov}
\usage{
addARTCov(dat, Args, oppSex = FALSE)
}
\arguments{
\item{dat}{Dataset to add ART vars to.}

\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{oppSex}{Make opposite-sex HIV prevalence. Default is FALSE.}
}
\value{
data.frame
}
\description{
Add community ART coverage.
}
\keyword{internal}
