% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircumcisionData.R
\name{addCircumcisionData}
\alias{addCircumcisionData}
\title{addCircumcisionData}
\usage{
addCircumcisionData(dat, cdat, mergeBy = c("IIntID", "Year"), Args)
}
\arguments{
\item{dat}{dataset with episodes.}

\item{cdat}{circumcision dataset.}

\item{mergeBy}{merge by IIntID or IIntID/Year.}
}
\description{
gets Circumcision data from MGH AHRI datasets.
}
