% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{addCircumVar}
\alias{addCircumVar}
\title{Adds the ever circumcised variable to an existing dataset.}
\usage{
addCircumVar(dat, cdat = NULL)
}
\arguments{
\item{dat}{An existing dataset.}

\item{cdat}{Circumcision dataset from \code{\link{getCircumcisionData}}.}
}
\value{
data.frame
}
\description{
Adds the ever circumcised variable to an existing dataset.
}
