% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{addHIVPrevBS}
\alias{addHIVPrevBS}
\title{addHIVPrevBS}
\usage{
addHIVPrevBS(inFile, dat, Args, Type = "All")
}
\arguments{
\item{inFile}{The filepath to the dataset with HIV prevalence.}

\item{dat}{A dataset to add HIV prevalence variables to.}

\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{Type}{Males, Females or All for ART coverage.}
}
\value{
data.frame
}
\description{
Calculates the HIV prevalence of area surrounding BS
}
