% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{addMigrVars}
\alias{addMigrVars}
\title{addMigrVars}
\usage{
addMigrVars(dat, mdat, carry = TRUE)
}
\arguments{
\item{dat}{Existing dataset to merge variables into.}

\item{mdat}{Dataset of migration variables \code{\link{makeMigrVars}}.}

\item{carry}{Whether to carry missing values forward and backward after merging. Default
is TRUE.}
}
\value{
data.frame
}
\description{
Adds variables from \code{\link{makeMigrVars}} to an
existing dataset.
}
\examples{
\donttest{
Args <- setArgs()
epi <- setEpisodes(Args) 
mdat <- makeMigrVars(Args)
epi2 <- addMigrVars(epi, mdat)
}
}
\keyword{internal}
