% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{AggByAge}
\alias{AggByAge}
\title{AggByAge}
\usage{
AggByAge(dat)
}
\arguments{
\item{dat}{A dataframe generated from \code{\link{getIncData}}.}
}
\value{
data.frame
}
\description{
Aggregates data to get sero events and pyears by age category.
}
\examples{
# Show for one imputation 
getFiles <- setFiles('/home/alain/Seafile/AHRI_Data/2020')
Args <- setArgs(imputeMethod = imputeMidPoint)
rtdat <- getRTData(dat = getHIV())
idat <- getIncData(rtdat, Args = Args)
inc <- AggByAge(idat)
}
