% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{AggByYear}
\alias{AggByYear}
\title{AggByYear}
\usage{
AggByYear(dat)
}
\arguments{
\item{dat}{An existing dataset.}
}
\value{
data.frame
}
\description{
Aggregates data to get sero events and pyears by year.
formula.
}
\examples{
# Show for one imputation 
Args <- setArgs(nSim=1, imputeMethod=imputeEndPoint)
getFiles <- setFiles(file.path(Sys.getenv("HOME"), "AHRI_Data/2019"))
hiv   <- getHIV()
rtdat <- getRTData(hiv)
bdat <- getBirthDate()
ydat <- getIncData(rtdat, bdat, Args)
inc <- AggByYear(ydat)
}
