% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{aggregateInc}
\alias{aggregateInc}
\title{aggregateInc}
\usage{
aggregateInc(dat)
}
\arguments{
\item{dat}{dataset}
}
\value{
data.frame
}
\description{
Gets the aggregated sero_event counts and total person-years by Year, Age,
and sex.
}
\examples{
edat <- splitAtEarlyPos(rtdat)
adat <- getAgeData(edat, Args)
inc <- aggregateInc(adat)
}
