% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahri.R
\docType{package}
\name{ahri}
\alias{ahri}
\title{ahri: R package for the AHRI datasets}
\section{The ahri package provides functions for analyzing the AHRI datasets.}{
NA
}

\section{Default .dta datasets}{

The ahri package uses the following deafult datasets, which you must request from the AHRI data
centre.
}

\section{setFile function}{

You can set the filepaths to the default .dta datasets using the \code{\link{setFiles}} function.
}

\seealso{
https://github.com/vando026/ahri
}
