% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{BSMax}
\alias{BSMax}
\title{BSMax}
\usage{
BSMax(inFile = Args$inFiles$demfile, outFile = "MaxBSIntID.csv")
}
\arguments{
\item{inFile}{file path to Demography dataset.}

\item{outFile}{file path to write dataset.}
}
\value{
data.frame
}
\description{
Gets the BSIntID that IIntID spent most time in a surveillance year.
}
\examples{
hiv <- BSMax(Args$inFiles$demfile)
}
