% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{calcARTCov}
\alias{calcARTCov}
\title{calcARTCov}
\usage{
calcARTCov(f = identity, cutoff = 13)
}
\arguments{
\item{f}{Use function to perform further operation on data, typically 
\code{\link{setData}}. Default is \code{identity}.}
}
\value{
data.frame
}
\description{
Calculate ART coverage for AHRI data.
}
