% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMortality.R
\name{calcMortality}
\alias{calcMortality}
\title{calcMortality}
\usage{
calcMortality(dat)
}
\arguments{
\item{dat}{a dataset}
}
\value{
table
}
\description{
Calculates annual mortality rates from \code{\link{getMortalityData}}.
}
\keyword{internal}
