% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcPoisCI}
\alias{calcPoisCI}
\title{calcPoisCI}
\usage{
calcPoisCI(dat, func = exp, by100 = TRUE, pval = FALSE)
}
\arguments{
\item{dat}{Estimates from glm, see for example \code{\link{doPoisYear}}.}

\item{func}{A function to transform the data, default is to exponentiate, \code{exp}.}

\item{by100}{If TRUE (default) calculate incidence per 100 person-years.}

\item{pval}{If TRUE (default) calculate the p-values.}
}
\value{
data.frame
}
\description{
Calculates standard errors for Poisson estimates from a single imputed
dataset.
}
\examples{
Args <- setArgs(Years=c(2008:2018), 
  Age=list(All=c(15, 45)),
  imputeMethod=imputeRandomPoint)
age_dat <- getAgeYear(dat=setHIV(Args))
rtdat <- getRTData(dat=getHIV())
idat <- getIncData(rtdat, bdat=getBirthDate(), Args)
pois_yr <- doPoisYear(idat, age_dat)
calcPoisCI(pois_yr)
}
