% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcPoisExact}
\alias{calcPoisExact}
\title{calcPoisExact}
\usage{
calcPoisExact(dat, byVar = "Year", fmt = TRUE)
}
\arguments{
\item{dat}{A dataset from \code{\link{AggByYear}} or \code{\link{AggByAge}}.}

\item{byVar}{Either "AgeCat" or "Year".}

\item{fmt}{If TRUE, format by 100 person-years and round off to three decimal places.}
}
\value{
data.frame
}
\description{
Calculate crude incidence rates using the Poisson exact method.
}
\examples{
Args <- setArgs(Years=c(2008:2018), 
  Age=list(All=c(15, 45)),
  imputeMethod=imputeRandomPoint)
hiv <- getHIV()
rtdat <- getRTData(hiv)
idat <- getIncData(rtdat, bdat=getBirthDate(), Args)
idat_yr <- AggByYear(idat)
calcPoisExact(idat_yr, byVar="Year")
}
