% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrend.R
\name{calcTrendYear}
\alias{calcTrendYear}
\title{calcTrendYear}
\usage{
calcTrendYear(Formula = "", dat)
}
\arguments{
\item{dat}{A dataset.}

\item{RHS}{The right hand side variable, must be string name.}
}
\value{
data.frame
}
\description{
Simpler fuction to calculate trends by Year.
}
\keyword{internal}
