% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{combineEst}
\alias{combineEst}
\title{combineEst}
\usage{
combineEst(dat)
}
\arguments{
\item{dat}{takes results from a function (i.e., \code{calcInc}.)}
}
\value{
list
}
\description{
Collect all estimates into single matrix
}
