% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{doPoisCrude}
\alias{doPoisCrude}
\title{doPoisCrude}
\usage{
doPoisCrude(dat)
}
\arguments{
\item{dat}{Dataset from \code{\link{AggByYear}}.}
}
\value{
data.frame
}
\description{
Calculates crude incidence rates by year using poisson regression.
}
