% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{doPoisYear}
\alias{doPoisYear}
\title{doPoisYear}
\usage{
doPoisYear(dat)
}
\arguments{
\item{dat}{Dataset from a function \code{\link{getIncData}}.}
}
\value{
data.frame
}
\description{
Do poisson regression for incidence rates by year or age.
}
\examples{
doPoisYear <- poisFunc("Year")
pdat <- getIncData(dat) 
doPoisYear(pdat)
doPoisAge <- poisFunc("AgeCat")
}
