% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{dropTasPData}
\alias{dropTasPData}
\title{dropTasPData}
\usage{
dropTasPData(dat)
}
\arguments{
\item{dat}{A dataset, which will be merged with the Bounded Structures
dataset, to determine if observations come from the TasP (northern) areas. 
If an observation cannot be linked to an area, it is kept.}
}
\value{
data.frame
}
\description{
Function to drop obervations from TasP areas.
}
