% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdat}
\alias{gdat}
\title{Example datasets for G-imputation: idat}
\format{
A data frame with 6042 rows and 6 variables:
\describe{
\item{IIntID}{The ID variable, name is compulsory}
\item{Time}{Time since the earliest-negative data}
\item{sero_date}{Indicates if repeat-tester seroconverted during observation period}
\item{Age0}{Mean-centered age variable}
\item{Age2}{Age0 squared}
\item{EverCircum}{EverCircumcised}
}
}
\usage{
gdat
}
\description{
A dataset consisting of the covariate values that will be input for the uniReg
function.
}
\keyword{datasets}
