% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeData.R
\name{getAgeData}
\alias{getAgeData}
\title{getAgeData}
\usage{
getAgeData(dat, idat, Args = eval.parent(quote(Args)))
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{idat}{individuals dataset from \code{\link{getFiles}}.}

\item{Args}{takes a list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
gets Age in a given surveillance year
}
\examples{
rtdata <- getRTData(hiv)
sdat <- splitRTData(rtdata)
adat <- getAgeData(sdata, idat)
}
