% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getAgeYear}
\alias{getAgeYear}
\title{getAgeYear}
\usage{
getAgeYear(dat)
}
\arguments{
\item{dat}{A dataset with an Age and Year variable.}
}
\value{
data.frame
}
\description{
Calculate mean age by year. This is mostly used in the \code{predict}
part of \code{\link{doPoisYear}}.
}
\examples{
Args <- setArgs(Years=c(2008:2018), 
  Age=list(All=c(15, 45)),
  imputeMethod=imputeRandomPoint)
rtdat <- getRTData(dat=getHIV())
idat <- getIncData(rtdat, bdat=getBirthDate(), Args)
age_dat <- getAgeYear(dat=setHIV(Args))
pois_yr <- doPoisYear(idat, age_dat)
}
