% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTCov.R
\name{getART}
\alias{getART}
\title{getART}
\usage{
getART(inFile = Args$inFiles$artemis)
}
\arguments{
\item{inFile}{path to data which is typically set using \code{\link{inFiles}}}
}
\value{
data.frame
}
\description{
get ARTemis data, mainly for calculating ART coverage
}
