% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{getARTDates}
\alias{getARTDates}
\title{getARTDates}
\usage{
getARTDates(dat)
}
\arguments{
\item{dat}{A dataset from \code{\link{getEpisodes}} which has the ART variables}
}
\value{
data.frame
}
\description{
Get ART initiation dates
}
\examples{
getARTDates()
}
