% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{getBirthDate}
\alias{getBirthDate}
\title{Gets the birthdates form the Surveillance Episodes dataframe.}
\usage{
getBirthDate(dat = NULL, addVars = " ")
}
\arguments{
\item{dat}{default value is \code{Null}, which calls \code{\link{getEpisodes}}}

\item{addVars}{String for a regular expression to select additional vars}
}
\value{
data.frame
}
\description{
Gets birth dates from \code{\link{getEpisodes}} data
}
\examples{
getBirthDate(addVars="Female")
}
