% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{getBSCord}
\alias{getBSCord}
\title{getBSCord}
\usage{
getBSCord(inFile = Args$inFile$bscfile)
}
\arguments{
\item{inFile}{path to the BSIntID coordinates csv file}
}
\value{
data.frame
}
\description{
get the GPS coordinates for BSIntID
}
