% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{getBSMax}
\alias{getBSMax}
\title{Gets the BSIntID that IIntID spent most time in a surveillance year.}
\usage{
getBSMax(dat = getEpisodes(), minDays = 0)
}
\arguments{
\item{dat}{The default is to call \code{\link{getEpisodes}}.}

\item{minDays}{Value of 1:366 min days spent in DSA to qualify as being a resident in
that year.}
}
\value{
data.frame
}
\description{
Gets the BSIntID that IIntID spent most time in a surveillance year.
}
\examples{
edat <- getEpisodes()
bs_max <- getBSMax(edat)
}
