% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircumcision.R
\name{getCircum}
\alias{getCircum}
\title{getCircum}
\usage{
getCircum(Keep)
}
\arguments{
\item{Keep}{keeps or drops circumcised men.}

\item{dat}{dataset with episodes.}

\item{Args}{arguments.}
}
\description{
gets Circumcision data from MGH AHRI datasets.
}
\examples{
keepCircum <- getCircum(Keep=1)
dropCircum <- getCircum(Keep=0)
keepCircum(dat, Args)
getCircumcision <- getCircum(Keep = c(0, 1))
}
