% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircumcisionData.R
\name{getCircumcisionData}
\alias{getCircumcisionData}
\title{getCircumcisionData}
\usage{
getCircumcisionData(inFile = Args$inFiles$mghfile)
}
\arguments{
\item{inFile}{arguments from \code{\link{setArgs}} to pass to \code{getHIV} function}
}
\description{
gets Circumcision data from MGH AHRI datasets. IsCircumcised: 1=yes;
2=no. WhenCircumcised: 1=as_child,2=as_adult,3=as_infant,>4 = NA.
}
