% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getCondomUseData}
\alias{getCondomUseData}
\title{getCondomUseData}
\usage{
getCondomUseData()
}
\description{
gets Condom use data from AHRI datasets. File path must be set in
\code{\link{setFiles}}.
}
