% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getCondomUseVar}
\alias{getCondomUseVar}
\title{Get variable for ever used condom status from Men and Women's health datasets.}
\usage{
getCondomUseVar()
}
\description{
Gets the ever used condom variable from Men and Women's health datasets.
}
