% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getCrudeMI}
\alias{getCrudeMI}
\title{getCrudeMI}
\usage{
getCrudeMI(dat)
}
\value{
data.frame
}
\description{
Calculates the rate and var according to code in
epitools::ageadjust.direct
}
