% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getCrudeRate}
\alias{getCrudeRate}
\title{getCrudeRate}
\usage{
getCrudeRate(dat)
}
\arguments{
\item{dat}{Dataset from \code{\link{combineEst}}.}
}
\value{
data.frame
}
\description{
Calculate crude incidence rate estimates using the standard formula.
}
\examples{
Args <- setArgs(nSim=10)
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
bdat <- getBirthDate(Args$inFiles$epifile)
dat <- mclapply(seq(Args$nSim), 
  function(i) calcInc(rtdat, bdat, Args))
cdat <- combineEst(dat)
getCrudeRate(cdat)
}
