% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVLData.R
\name{getCVLData}
\alias{getCVLData}
\title{getCVLData}
\usage{
getCVLData(inFile = Args$inFiles$pvlfile, Args)
}
\arguments{
\item{inFile}{path to the Stata .dta file}

\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Pulls in Community Viral Load data
}
