% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTData.R
\name{getDates}
\alias{getDates}
\title{getDates}
\usage{
getDates(dat, f)
}
\arguments{
\item{Var}{a variable name.}

\item{Name}{new variable name.}
}
\value{
data.frame
}
\description{
Function to get earliest and latest test dates
}
