% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{getDatesMin}
\alias{getDatesMin}
\title{Function to get earliest HIV test dates.}
\usage{
getDatesMin(dat, Var, Name)
}
\arguments{
\item{dat}{a dataset.}

\item{Var}{a variable name.}

\item{Name}{new variable name.}
}
\value{
data.frame
}
\description{
Function to get earliest test dates
}
