% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeath.R
\name{getDeaths}
\alias{getDeaths}
\title{getDeaths}
\usage{
getDeaths(inFile = NULL, dat)
}
\arguments{
\item{dat}{dataset from \code{\link{splitData}}.}
}
\value{
data.frame
}
\description{
Gets mortality data from Demography dataset.
}
