% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemographyData.R
\name{getDemographyData}
\alias{getDemographyData}
\title{getDemographyData}
\usage{
getDemographyData(inFile = Args$inFiles$demfile)
}
\arguments{
\item{inFile}{path to data, see \code{\link{getFiles}}.}
}
\value{
data.frame
}
\description{
get AHRI demography dataset
}
