% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{getEpisodes}
\alias{getEpisodes}
\title{getEpisodes}
\usage{
getEpisodes(inFile = NULL)
}
\arguments{
\item{inFile}{File path to the dataset, default is set to \code{getFiles()$epi_rda}.}
}
\value{
data.frame
}
\description{
Loads Episodes .Rda into memory, see \code{\link{readEpisodes}}.
}
