% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getEstimates}
\alias{getEstimates}
\title{getEstimates}
\usage{
getEstimates(dat, Args, By = "Year")
}
\arguments{
\item{dat}{takes dataset from a function (i.e., \code{doIncData}.)}

\item{Args}{takes list from \code{\link{setArgs}}.}

\item{By}{calculate by Year or AgeCat.}
}
\value{
data.frame
}
\description{
Once dates are imputed get the estimates.
}
\examples{
getIncidence <- function(Args) {
  hiv   <- getHIV(Args)
  rtdat <- getRTData(hiv)
  dat <- lapply(seq(Args$nSimulations),
    function(i) doIncData(rtdat, Args))
  out <- getEstimates(dat, Args) 
  out
}
}
