% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getEstMI}
\alias{getEstMI}
\title{getEstMI}
\usage{
getEstMI(dat, fun, Args = eval.parent(quote(Args)),
  calcBy = eval.parent(quote(By)))
}
\value{
data.frame
}
\description{
Calculates the standard errors for multiple imputation following formula given in P. Allison Missing Data
book, pg 30, in '~/Dropbox/Textbooks/Statistics'.
}
