% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getEstSI}
\alias{getEstSI}
\title{getEstSI}
\usage{
getEstSI(dat, fun, calcBy = eval.parent(quote(By)))
}
\value{
data.frame
}
\description{
Calculates the rate and confidence intervals for single imputation such as mid-point or
end-point.
}
