% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{getEverART}
\alias{getEverART}
\title{getEverART}
\usage{
getEverART(dat = getEpisodes())
}
\arguments{
\item{dat}{A dataset from \code{\link{getEpisodes}}.}
}
\value{
data.frame
}
\description{
Creates an Ever on ART variable.
}
\examples{
getEverART()
}
