% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{getFiles}
\usage{
getFiles(root = file.path(Sys.getenv("USERPROFILE"),
  "Documents/AC_Data/Derived"))
}
\arguments{
\item{root}{file path to derived files}
}
\value{
list
}
\description{
get/set file paths to AHRI data
}
\examples{
inFiles <- getFiles()
}
