% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{getHIV}
\alias{getHIV}
\title{getHIV}
\usage{
getHIV(inFile = getFiles()$hiv_rda)
}
\arguments{
\item{inFile}{File path from \code{\link{setFiles}}.}
}
\value{
data.frame
}
\description{
Get all valid test results >15 yrs age from HIV surveillance.
}
