% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVContact}
\alias{getHIVContact}
\title{getHIVContact}
\usage{
getHIVContact(Args, dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dat}{Dataset, otherwise made with \code{\link{setHIVMiss}}.}
}
\value{

}
\description{
Get participants who are eligible and contacted for HIV test.
}
