% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTData.R
\name{getHIVDatesLong}
\alias{getHIVDatesLong}
\title{getHIVDatesLong}
\usage{
getHIVDatesLong(dat)
}
\arguments{
\item{dat}{Dateset from \code{\link{getHIV}}.}
}
\value{

}
\description{
Gets HIV dates and saves in long format. Used mainly for IntCens.
}
