% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVEligible}
\alias{getHIVEligible}
\title{getHIVEligible}
\usage{
getHIVEligible(dat)
}
\arguments{
\item{dat}{Dataset from \code{\link{setHIVMiss}}.}
}
\value{

}
\description{
Get eligibility for HIV testing.
}
