% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVMiss2}
\alias{getHIVMiss2}
\title{getHIVMiss2}
\usage{
getHIVMiss2(Root = setRoot())
}
\arguments{
\item{getFiles}{}
}
\value{

}
\description{
Gets missed test dates but pulls each HIV surveillance dataset by year.
}
