% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVRefused}
\alias{getHIVRefused}
\title{getHIVRefused}
\usage{
getHIVRefused(Args, dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dat}{Dataset, otherwise made with \code{\link{setHIVMiss}}.}
}
\value{
data.frame
}
\description{
Get summary of participants that refused to test.
}
