% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncData}
\alias{getIncData}
\title{getIncData}
\usage{
getIncData(rtdat, idat, Args)
}
\arguments{
\item{rtdat}{dataset from \code{\link{getRTData}}.}

\item{idat}{dataset from \code{\link{getBirthData}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Function used to prepare the data for \code{\link{getIncidence}}.
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
idat <- getBirthDate(Args$inFiles$epifile)
dat <- Args$imputeMethod(rtdat)
edat <- splitAtSeroDate(dat, splitYears=Args$Years) 
adat <- getAgeData(edat, idat,  Args)
getIncData(rtdat, idat, Args)
}
