% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(Args = setArgs(), formulas = getFormula())
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}.}

\item{formulas}{A list of formulas for the poisson regression models, see
\code{\link{getFormula}}.}
}
\value{
list
}
\description{
Calculate annual HIV incidence rates using the single random-point
approach and multiple imputation.
}
