% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{getIntCensDates}
\alias{getIntCensDates}
\title{getIntCensDates}
\usage{
getIntCensDates(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Make HIV test dates for IntCens model
}
