% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMortality.R
\name{getMortalityData}
\alias{getMortalityData}
\title{getMortalityData}
\usage{
getMortalityData(Args, startVar = "HIVPositive")
}
\arguments{
\item{Args}{see \code{\link{setArgs}}.}

\item{startDate}{string variable at which person-time starts. Use early_pos
for HIV-positive all-cause mortality, otherwise obs_start (earliest HIV-neg) for all-cause
mortality.}
}
\value{
data.frame
}
\description{
gets mortality data.
}
