% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartnerData.R
\name{getPartnerData}
\alias{getPartnerData}
\title{getPartnerData}
\usage{
getPartnerData(Args)
}
\arguments{
\item{inFile}{path to data which is typically set using \code{\link{setFiles}}}
}
\value{
data.frame
}
\description{
get data on partners and marital status.
}
\keyword{internal}
