% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrend.R
\name{getPopAgg}
\alias{getPopAgg}
\title{getPopAgg}
\usage{
getPopAgg(dat, Formula = "IIntID ~ Year")
}
\arguments{
\item{Args.}{}

\item{Formula.}{}
}
\value{
list
}
\description{
Gets aggregated numbers by category.
}
\examples{
Args <- setArgs()
hiv <- getHIV(Args)
hiv <- setData(hiv)
getPopAgg(hiv, "IIntID ~ Year + HIVResult")
}
