% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVLData.R
\name{getPVLData}
\alias{getPVLData}
\title{getPVLData}
\usage{
getPVLData(cvl, Args)
}
\arguments{
\item{cvl}{dataset from \code{\link{getVLData}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Makes the PVL version of viral load data for 2011, 2013, 2014.
}
