% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTData.R
\name{getRTData}
\alias{getRTData}
\title{getRTData}
\usage{
getRTData(dat, Args = eval.parent(quote(Args)))
}
\arguments{
\item{dat}{dataset from \code{\link{getHIV}}.}

\item{Args}{takes an Args list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Get all repeat testers from HIV surveillance.
}
\examples{
hiv <- getHIV(Args)
rtdat <- getRTData(hiv)
}
