% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVLData.R
\name{getVLData}
\alias{getVLData}
\title{getVLData}
\usage{
getVLData(Args)
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Gets viral load data for 2011, 2013, 2014.
}
