% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{getWeights}
\usage{
getWeights(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\description{
Get weights for age and sex adjustement.
}
