% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getWGH}
\alias{getWGH}
\title{getWGH}
\usage{
getWGH(inFile = NULL)
}
\arguments{
\item{inFile}{Filepath to .rda dataset, default is \code{getFiles()$wgh_rda}. Leave as
NULL if you don't know what to do or see \code{\link{setFiles}}.}
}
\value{
data.frame
}
\description{
Reads in Women's general health data.
}
