% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{getYearDates}
\alias{getYearDates}
\title{getYearDates}
\usage{
getYearDates(Years)
}
\arguments{
\item{Years}{The years by which to split the data.}
}
\value{
vector
}
\description{
gets numeric dates for 31DecYYYY.
}
