% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMethods.R
\name{imputeEndPoint}
\alias{imputeEndPoint}
\title{imputeEndPoint}
\usage{
imputeEndPoint(dat)
}
\arguments{
\item{dat}{dataset from \code{\link{getRTData}}.}
}
\value{
data.frame
}
\description{
Impute a end-point seroconversion date within the censored interval.
}
