% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMethods.R
\name{imputeMidPoint}
\alias{imputeMidPoint}
\title{Impute a mid-point seroconversion date within the censored interval.}
\usage{
imputeMidPoint(dat)
}
\arguments{
\item{dat}{dataset from \code{\link{getRTData}}.}
}
\value{
data.frame
}
\description{
Impute a mid-point seroconversion date within the censored interval.
}
