% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMethods.R
\name{imputeRandomPoint}
\alias{imputeRandomPoint}
\title{imputeRandomPoint}
\usage{
imputeRandomPoint(dat)
}
\arguments{
\item{dat}{dataset from \code{\link{getRTData()}}.}
}
\value{
data.frame
}
\description{
Impute random seroconversion date(s) within the censored interval.
}
