% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{incTab}
\alias{incTab}
\title{incTab}
\usage{
incTab(obj)
}
\arguments{
\item{obj}{takes dataset.}

\item{age}{Do for age or year.}
}
\value{
data.frame
}
\description{
Make table for excel.
}
