% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{makeAgeVars}
\alias{makeAgeVars}
\title{makeAgeVars}
\usage{
makeAgeVars(dat, visitdate = NULL, age_cut = NULL)
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{visitdate}{Variable name as string for Date of visit.}

\item{age_cut}{Vector of ages to make age categories, if NULL then mean centered age
and age-squared variables made.}
}
\value{
data.frame
}
\description{
Function for making age variables
}
\examples{
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
}
