% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{makeMigrVars}
\alias{makeMigrVars}
\title{makeMigrVars}
\usage{
makeMigrVars(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Make migration variables: the cumulative time spent out the PIP
study area and the number of migration events (in and out) by year.
}
