% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartnerData.R
\name{makePartnerData}
\alias{makePartnerData}
\title{makePartnerData}
\usage{
makePartnerData(dat, Args = eval.parent(quote(Args)))
}
\arguments{
\item{dat}{data.frame to which partner data is merged.}

\item{inFile}{path to data which is typically set using \code{\link{inFiles}}}
}
\value{
data.frame
}
\description{
Adds partner data from \code{\link{getPartnerData}} to existing dataset.
Missing values are inputed or carried forward.
}
