% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{makePropRes}
\alias{makePropRes}
\title{Makes a variable of the proportion of time that a participant spent
as a resident in the PIP surveillance area by year.}
\usage{
makePropRes(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Makes a variable of the proportion of time that a participant spent as a
resident in the PIP surveillance arear by year.
}
\examples{
dat <- setEpisodes(setArgs()) 
adat <- makePropRes(setArgs())
dat <- dplyr::left_join(dat, adat, by=c("IIntID", "Year"))
dplyr::select(dat, IIntID, Year, ExpDays, Resident, PropRes)
}
