% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{MIaggregate}
\alias{MIaggregate}
\title{MIaggregate}
\usage{
MIaggregate(dat, col_names = c("sero_event", "pyears"))
}
\arguments{
\item{dat}{List of results.}

\item{col_names}{Names used to collect columns into a data.frame.}
}
\value{
data.frame
}
\description{
Aggregates results after mutiple imputation.
}
\examples{
Args <- setArgs(nSim=2, mcores=1)
rtdat <- getRTData(dat=getHIV())
mdat <- MIdata(rtdat, Args)
mdat <- mitools::imputationList(mdat)
inc <- with(mdat, fun=AggByYear)
MIaggregate(inc)
}
