% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{MIpredict}
\alias{MIpredict}
\title{MIpredict}
\usage{
MIpredict(object, newdata)
}
\arguments{
\item{object}{Results from the object generated by \code{MIcombine}.}

\item{newdata}{Dataframe of variables for prediction.}
}
\value{
data.frame
}
\description{
Used with \code{mitools} to get incidence rate estimates per 100
person-years after multiple imputation.
}
\examples{
rtdat <- getRTData(dat=getHIV())
Args <- setArgs(nSim=2, Years=c(2008:2018))
mdat <- MIdata(rtdat, Args)
mdat <- mitools::imputationList(mdat)
F1 <- "sero_event ~ -1 + as.factor(Year) + Age +
  as.factor(Year):Age + offset(log(tscale))" 
mods <- with(mdat, glm(as.formula(F1), family=poisson))
betas <- mitools::MIextract(mods,fun=coef)
var <- mitools::MIextract(mods, fun=vcov)
res <-  mitools::MIcombine(betas, var) 
MIpredict(object=mods, newdata=getAgeYear(setHIV(Args)))
}
