% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMax.R
\name{mkBSData}
\alias{mkBSData}
\title{mkBSData}
\usage{
mkBSData(inFiles)
}
\arguments{
\item{inFile}{file path to import Demography dataset.}

\item{outFile}{file path to write dataset.}
}
\value{
data.frame
}
\description{
Gets the BSIntID that IIntID spent most time in a surveillance year.
}
