% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotIncAge}
\alias{plotIncAge}
\title{plotIncAge}
\usage{
plotIncAge(age_mal, age_fem, Colors = c("grey50", "grey80"),
  fname = "age_plot", TIFF = TRUE)
}
\arguments{
\item{dat}{dataset from \code{\link{getIncidence()}}.}
}
\description{
Plot HIV incidence by Age.
}
