% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotIncPrev}
\alias{plotIncPrev}
\title{plotIncPrev}
\usage{
plotIncPrev(inc, prev = NULL, Args = NULL, gfun = png, bwidth = 2,
  fname = "test")
}
\arguments{
\item{inc}{Incidence data.}

\item{prev}{Prevalence data.}
}
\description{
Plots incidence and prevalence.
}
