% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotIncSex}
\alias{plotIncSex}
\title{plotIncSex}
\usage{
plotIncSex(
  Mal,
  Fem,
  yLim = 7,
  Colors = c("blue", "red"),
  gfun = png,
  Title = "",
  fname = "year_plot"
)
}
\arguments{
\item{Mal}{data.frame with rate, lci, and uci for men.}

\item{Fem}{data.frame with rate, lci, and uci for women.}
}
\description{
Plot HIV incidence with 95% CIs as whiskers.
}
\keyword{internal}
