% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotIncSexArea}
\alias{plotIncSexArea}
\title{plotIncSexArea}
\usage{
plotIncSexArea(
  Mal,
  Fem,
  yLim = 5.5,
  Colors = c(Blues[3], Reds[3]),
  gcolor = "grey50",
  ipoints = TRUE,
  bwidth = list(mal = c(2.5, 2.5), fem = c(2.5, 2.5)),
  Legend = c("Men", "Women"),
  gfun = png,
  Title = "",
  fname = "year_plot"
)
}
\arguments{
\item{Mal}{data.frame with rate, lci, and uci for men.}

\item{Fem}{data.frame with rate, lci, and uci for women.}
}
\description{
Plot HIV incidence with 95% CIs as area.
}
\keyword{internal}
