% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotKSInc}
\alias{plotKSInc}
\title{plotKSInc}
\usage{
plotKSInc(inc1, inc2 = NULL, inc3 = NULL, Colors = c("grey60", "grey70",
  NULL), bwidth = c(2, 2, NULL), ylim1 = c(0, 7), gcolor = "grey50",
  TIFF = TRUE, Legend = c("Men", "Women", NULL), title = "",
  fname = "year_plot")
}
\arguments{
\item{dat}{dataset from \code{\link{getIncidence()}}.}
}
\description{
Plot HIV incidence.
}
