% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{readARTCov}
\alias{readARTCov}
\title{readARTCov}
\usage{
readARTCov(Female = 1)
}
\arguments{
\item{Female}{Read female data (Female = 1) or male data (Female = 0)}
}
\value{
data.frame
}
\description{
Read in community ART coverage data from Source path
}
\keyword{internal}
