% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{readBSData}
\alias{readBSData}
\title{readBSData}
\usage{
readBSData(inFile = getFiles()$bsifile)
}
\arguments{
\item{inFile}{filepath to data.}
}
\value{
data.frame
}
\description{
Read in Bounded Structures data.
}
\examples{
readBSData() 
}
