% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{readHealthData}
\alias{readHealthData}
\title{readHealthData}
\usage{
readHealthData(inFile, Fem)
}
\arguments{
\item{inFile}{Filepath to dataset, default is \code{getFiles()$mghfile}.}
}
\value{

}
\description{
Reads in men and women general health data.
}
\examples{
readHealthMen <- readHealthData(getFiles()$mghfile)
readHealthWomen <- readHealthData(getFiles()$wghfile)
}
