% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{readHealthDataMen}
\alias{readHealthDataMen}
\title{readHealthDataMen}
\usage{
readHealthDataMen()
}
\arguments{
\item{inFile}{Filepath to dataset, default is \code{getFiles()$mghfile}.}
}
\value{

}
\description{
Reads in men general health data.
}
\examples{
readHealthDataMen <- readHealthData(getFiles()$mghfile)
}
