% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{readHealthDataWomen}
\alias{readHealthDataWomen}
\title{readHealthDataWomen}
\usage{
readHealthDataWomen()
}
\arguments{
\item{inFile}{Filepath to dataset, default is \code{getFiles()$fghfile}.}
}
\value{

}
\description{
Reads in women general health data.
}
\examples{
readHealthWomen <- readHealthData(getFiles()$wghfile)
}
