% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(inFile = Args$inFiles$hiv_dta,
  outFile = Args$inFiles$hivfile)
}
\arguments{
\item{inFile}{path from \code{\link{setArgs}}.}

\item{outFile}{path from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
