% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{readHIVSurvYear}
\alias{readHIVSurvYear}
\title{readHIVSurvYear}
\usage{
readHIVSurvYear(inFile, addVars = " ")
}
\arguments{
\item{inFile}{File path to file.}
}
\value{
data.frame
}
\description{
Reads the data for each HIV surveillance dataset by year.
}
