% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{readHSEData}
\alias{readHSEData}
\title{readHSEData}
\usage{
readHSEData(inFile = getFiles()$hsefile)
}
\arguments{
\item{inFile.}{File path from \code{\link{setFiles}}.}
}
\value{
data.frame
}
\description{
Read HSE data.
}
