% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{readMGH}
\alias{readMGH}
\title{readMGH}
\usage{
readMGH(inFile = NULL, write_rda = TRUE)
}
\arguments{
\item{inFile}{Filepath to .dta dataset, default is \code{getFiles()$mghfile}.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads the Men (MGH) General Health Dataset.
}
\examples{
readMGH(write_rda = FALSE)
}
