% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{readPIPData}
\alias{readPIPData}
\title{readPIPData}
\usage{
readPIPData(inFile = getFiles()$pipfile)
}
\arguments{
\item{inFile}{}
}
\value{
data.frame
}
\description{
Read in PIP data to identify ACIDS from TASP area
}
