% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{readUniRegResults}
\alias{readUniRegResults}
\title{readUniRegResults}
\usage{
readUniRegResults(File = NULL)
}
\arguments{
\item{File}{File path to \code{\link{uniReg}} output.txt.}
}
\description{
Read the results from the \code{\link{uniReg}} model into .Rdata
format.
}
