% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{readWGH}
\alias{readWGH}
\title{Reads the Women's  General Health Dataset (WGH).}
\usage{
readWGH(inFile = NULL, write_rda = TRUE)
}
\arguments{
\item{inFile}{Filepath to .dta dataset, default is \code{getFiles()$wghfile}.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads the Women's  General Health Dataset (WGH).
}
\examples{
readWGH(write_rda = FALSE)
}
