% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtdat}
\alias{rtdat}
\title{Example datasets for G-imputation}
\format{
A data frame with 1022 rows and 6 variables:
\describe{
\item{IIntID}{The ID variable, which must have the name \code{IIntID}}
\item{Female}{If participant female, 1=yes, 0=no}
\item{obs_start}{The earliest-negative date}
\item{late_neg}{The latest-negative date}
\item{early_pos}{The earliest-positive date}
\item{sero_event}{Indicates if repeat-tester seroconverted during observation period}
}
}
\usage{
rtdat
}
\description{
The repeat-testers dataset, with one row per individual, consisting of the start date,
latest-negative and earliest-positive date. Your dataset must have these names to work
with the \code{ahri} functions.
}
\keyword{datasets}
