% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{saveInc}
\alias{saveInc}
\title{saveInc}
\usage{
saveInc(obj, out = output)
}
\arguments{
\item{obj}{takes object.}

\item{out}{File path to write.}
}
\description{
Save to .Rdata file.
}
