% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeData.R
\name{setAge}
\alias{setAge}
\title{setAge}
\usage{
setAge(dat, Args)
}
\arguments{
\item{dat}{must have a variable named AgeAtVisit}

\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
sets age according to arguments
}
