% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeData.R
\name{setAgeYear}
\alias{setAgeYear}
\title{setAgeYear}
\usage{
setAgeYear(dat, idat, Args = eval.parent(quote(Args)))
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{idat}{Individuals dataset (Date of Birth variable) from \code{\link{getBirthDate}}.}

\item{Args}{takes a list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
sets Age and survey year according to values in
\code{\link{setArgs}}.
}
\examples{
rtdata <- getRTData(hiv)
idat <- getBirthDate(Args$inFiles$epifile)
sdat <- splitAtEarlyPos(rtdat)
adat <- setAgeYear(sdat, idat)
}
