% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{setData}
\alias{setData}
\title{setData}
\usage{
setData(dat, bdat = getBirthDate(), Args = eval.parent(quote(Args)))
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{Args}{takes a list from \code{\link{setArgs}}.}

\item{idat}{Date of Birth variable from \code{\link{getBirthDate}}.}
}
\value{
data.frame
}
\description{
sets data according to values in \code{\link{setArgs}}.
}
\examples{
rtdata <- getRTData(hiv)
bdat <- getBirthDate(Args$inFiles$epifile)
sdat <- splitAtEarlyPos(rtdat)
adat <- setData(sdat, bdat)
}
