% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemographyData.R
\name{setDemography}
\alias{setDemography}
\title{setDemography}
\usage{
setDemography(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Makes an age variable and filters by age
}
