% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemographyData.R
\name{setDemographyData}
\alias{setDemographyData}
\title{setDemographyData}
\usage{
setDemographyData(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Sets the data according to Args.
}
\examples{
setDemographyData(Args)
}
