% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{setEpisodes}
\alias{setEpisodes}
\title{setEpisodes}
\usage{
setEpisodes(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
set episodes data according to Args.
}
\examples{
Args <- setArgs()
setEpisodes(Args)
}
