% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{setFiles}
\alias{setFiles}
\title{setFiles}
\usage{
setFiles(
  folder = "",
  hivfile = "RD05-99 ACDIS HIV All.dta",
  epifile = "SurveillanceEpisodesExtended.dta",
  wghfile = "RD03-99 ACDIS WGH ALL.dta",
  mghfile = "RD04-99 ACDIS MGH ALL.dta",
  bsifile = "RD01-03 ACDIS BoundedStructures.dta",
  hiv_rda = "ACDIS_HIV_All.Rda",
  epi_rda = "SurveillanceEpisodesExtended.Rda",
  wgh_rda = "ACDIS_WGH_ALL.Rda",
  mgh_rda = "ACDIS_MGH_ALL.Rda",
  bsc_rda = "ACDIS_BoundedStructures.Rda"
)
}
\arguments{
\item{folder}{The path (as a string) to  the folder of default .dta files. If
\code{folder=""}, then the default is to call a dialogue box for you to point and click
to the folder path.}

\item{hivfile}{Reads the file RD05-99 ACDIS HIV All.dta}

\item{epifile}{Reads the file SurveillanceEpisodesExtended.dta}

\item{wghfile}{Reads the file RD03-99 ACDIS WGH ALL.dta}

\item{mghfile}{Reads the file RD04-99 ACDIS MGH ALL.dta}

\item{bsifile}{Reads the file RD01-03 ACDIS BoundedStructures.dta}

\item{hiv_rda}{Writes #' the file ACDIS_HIV_All.Rda}

\item{epi_rda}{Writes the file SurveillanceEpisodesExtended.Rda}

\item{wgh_rda}{Writes the file ACDIS_WGH_ALL.Rda}

\item{mgh_rda}{Writes the file ACDIS_MGH_ALL.Rda}

\item{bsc_rda}{Writes the file ACDIS_BoundedStructures.Rda}
}
\value{
function
}
\description{
Set file paths to the default AHRI datasets, which must be placed into a
single folder. The names of the default .dta datasets must be exactly the same as
described below: The function also sets the names for the .Rda datasets used in other
functions.  You should change these names only if you know what you are doing (assume
that you don't). The \code{ahri} package will never overwrite the default. dta
datasets. You must assign the \code{setFiles} function to a name called
\code{getFiles}. See the examples below.
}
\examples{
# You must assign the setfiles function to the getFiles name 
getFiles <- setFiles(folder="Path/to/my/datafolder")

# If for some reason your HIV surveillance .dta file is named differently 
getFiles <-setFiles(folder="Path/to/my/datafolder", hivfile="RD09-01 PIP HIV All.dta")

# print out the file paths and names 
setFiles() 
}
