% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{setHIVMiss}
\alias{setHIVMiss}
\title{setHIVMiss}
\usage{
setHIVMiss(Root = setRoot(), dropTasP = TRUE)
}
\arguments{
\item{Root}{The root path to folder containing HIV surveillance datasets.}
}
\value{

}
\description{
Gets HIV test dates by merging yearly HIV surveillance datasets.
}
