% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{smoothInc}
\alias{smoothInc}
\title{smoothInc}
\usage{
smoothInc(dat, x = "time", y = "rate", bwidth = 1)
}
\arguments{
\item{dat}{takes dataset from \code{\link{getIncidence}}.}

\item{bwidth}{bandwith for \code{ksmooth} function.}
}
\value{
data.frame
}
\description{
get smoothed incidence estimates.
}
\examples{
smoothInc(dat, x="time", y="rate")
}
