% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitAtEarlyPos}
\alias{splitAtEarlyPos}
\title{splitAtEarlyPos}
\usage{
splitAtEarlyPos(dat = NULL, splitYears = NULL)
}
\arguments{
\item{dat}{a dataset}

\item{splitYears}{Vector from \code{\link{getYearDates}}.}
}
\value{
data.frame
}
\description{
Split data at the earliest HIV-positive date.
}
\examples{
Args <- setArgs(Years=c(2008:2019))
hiv <- setHIV(Args)
rtdat <- getRTData(hiv)
splitAtEarlyPos(rtdat, splitYears=Args$Years)
}
