% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitAtSeroDate}
\alias{splitAtSeroDate}
\title{Split data at the imputed seroconversion date.}
\usage{
splitAtSeroDate(dat = NULL, splitYears = NULL)
}
\arguments{
\item{dat}{a dataset}

\item{splitYears}{Vector from \code{\link{getYearDates}}.}
}
\value{
data.frame
}
\description{
Split data at the imputed seroconversion date.
}
\examples{
Args <- setArgs(Years=c(2008:2020))
hiv <- setHIV(Args)
rtdat <- getRTData(hiv)
sdat <- imputeMidPoint(rtdat)
splitAtSeroDate(sdat, splitYears=Args$Years)
}
