% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData2}
\alias{splitData2}
\title{splitData2}
\usage{
splitData2(dat, years = c(2003:2018))
}
\arguments{
\item{dat}{Dataset must have variables called obs_start, obs_end and event.}

\item{years}{The years by which to split the data.}
}
\value{
data.frame
}
\description{
Function for splitting data into episodes.
}
