% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{sumHIVMiss}
\alias{sumHIVMiss}
\title{sumHIVMiss}
\usage{
sumHIVMiss(dat)
}
\arguments{
\item{dat}{Dataset from \code{\link{getHIVEligible}}.}
}
\value{
data.frame
}
\description{
Get summary of participants that refused to test.
}
\keyword{internal}
