% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ahri.R
\name{test_ahri_dta}
\alias{test_ahri_dta}
\title{test_ahri_dta}
\usage{
test_ahri_dta(year = "2020", report = "progress")
}
\arguments{
\item{year}{The release year that you want to perform the tests on. Currently, tests
include 2019 and 2020 release years only.}

\item{report}{How to display results, can be "minimal", "summary", "progress".}
}
\description{
Tests to see if you are using the correct default .dta datasets. The
tests will generate summary values of selected variables from the datasets and check if they match
expected values defined in the test.
}
