% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{uniReg}
\alias{uniReg}
\title{uniReg}
\usage{
uniReg(
  InFile,
  OutFile,
  Model,
  ID = NULL,
  inf = "Inf",
  iter = 5000,
  cthresh = 1e-04,
  r = 1,
  printout = FALSE,
  ign_stout = TRUE
)
}
\arguments{
\item{InFile}{txt file to be input}

\item{OutFile}{txt file to be output}

\item{Model}{equation to be given}

\item{ID}{name of subject ID}

\item{inf}{Value for infinite, default is "Inf"}

\item{iter}{Number of iterations}

\item{cthresh}{Threshold for convergence}

\item{r}{Threshold for convergence}

\item{printout}{Print results to screen}

\item{ign_stout}{For Linux systems}
}
\description{
Wrapper for Intcens executable by Zeng et al 2016. See
http://dlin.web.unc.edu/software/intcens/ for details. The executable is shipped with 
the \code{ahri} package. Type in the R console \code{system.file("intcens", "unireg.exe", package =
"ahri")}.
}
\examples{
\donttest{
uniReg(
InFile = file.path(output, "TestSim.txt"),
OutFile = file.path(output, "TestSim_Out.txt"), r = 0.0,
Model = "(time, d) = v1 + v2 + v3",
ID = "id")}
}
