% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{addBSVars}
\alias{addBSVars}
\title{addBSVars}
\usage{
addBSVars(dat, Vars = "Area", dropMissBS = TRUE)
}
\arguments{
\item{dat}{An existing dataset.}

\item{Vars}{Select variables.}

\item{dropMissBS}{Drop any missing BS.}
}
\value{
data.frame
}
\description{
Add variables from Bounded Structures to existing dataset.
}
