% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{AggByAge}
\alias{AggByAge}
\title{AggByAge}
\usage{
AggByAge(dat)
}
\arguments{
\item{dat}{Dataset.}
}
\value{
data.frame
}
\description{
Aggregates data to get sero events and pyears by age category.
formula.
}
