% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{AggByYear}
\alias{AggByYear}
\title{AggByYear}
\usage{
AggByYear(dat)
}
\arguments{
\item{dat}{A dataframe generated from \code{\link{getIncData}}.}
}
\value{
data.frame
}
\description{
Aggregates data to get sero events and pyears by year.
}
\examples{
Args <- setArgs(imputeMethod = imputeMidPoint)
rtdat <- getRTData(dat = getHIV())
idat <- getIncData(rtdat, Args = Args)
inc <- AggByYear(idat)
}
