% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{AggFunc}
\alias{AggFunc}
\title{AggFunc}
\usage{
AggFunc(RHS)
}
\arguments{
\item{RHS}{The variable name as string that you want to aggregate by.}
}
\description{
A function factory for creating specific AggByFuncs, see for example \code{\link{AggByYear}}.
}
\examples{
\donttest{
AggByYear <- AggFunc("Year")
}
}
