% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMax.R
\name{BSMax}
\alias{BSMax}
\title{BSMax}
\usage{
BSMax(inFile = file.path(Sys.getenv("USERPROFILE"),
  "Documents/AC_Data/Derived/Demography/2016",
  "RD02-01_ACDIS_Demography.csv"), outFile = "MaxBSIntID.csv")
}
\arguments{
\item{inFile}{file path to import Demography dataset.}

\item{outFile}{file path to write dataset.}
}
\value{
data.frame
}
\description{
Gets the BSIntID that IIntID spent most time in a surveillance year.
}
\examples{
hiv <- getIncData(edat)
}
