% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{calcARTCov}
\alias{calcARTCov}
\title{calcARTCov}
\usage{
calcARTCov(dat = getEverART(), Args = setArgs())
}
\arguments{
\item{dat}{A dataset from \code{getEverART}.}

\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Calculate ART coverage for AHRI data. (This is a very crude measure of
ART coverage. More work needed on an appropriate measure.)
}
