% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{calcARTCov}
\alias{calcARTCov}
\title{calcARTCov}
\usage{
calcARTCov(dat = getEverART(), Args = setArgs(), Formula = "EverART ~ Year")
}
\arguments{
\item{dat}{A dataset from \code{\link{getEverART}}.}

\item{Args}{requires Args, see \code{\link{setArgs}}}

\item{Formula}{A formula using R formula language to get ART estimates by group,
default is by year \code{"EvertART ~ Year"}}
}
\value{
data.frame
}
\description{
Calculate ART coverage for AHRI data. (This is a very crude measure of
ART coverage. More work needed on an appropriate measure.)
}
\examples{
\donttest{
calcARTCov(Formula = "EverART ~ Year + Female + AgeCat")
}
}
