% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{calcARTCov}
\alias{calcARTCov}
\title{calcARTCov}
\usage{
calcARTCov(Args, cutoff = 13)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}}

\item{cutoff}{If after month, ART for that year doesn't count, default is 13 for ART in
any month.}
}
\value{
data.frame
}
\description{
Calculate ART coverage for AHRI data.
}
\keyword{internal}
