% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcInc}
\alias{calcInc}
\title{calcInc}
\usage{
calcInc(dat, fun, calcBy = "Year")
}
\arguments{
\item{dat}{Dataset from \code{\link{aggregateInc}}.}

\item{fun}{Function to use for calculations}

\item{calcBy}{Results by Year, Age, or Sex.}
}
\value{
data.frame
}
\description{
Calculates the crude and adjusted incidence.
}
\examples{
edat <- censorData(rtdat,Args)
adat <- getAgeData(edat, Args)
inc <- aggregateInc(adat)
calcInc(inc, getCrudeMI, calcBy="Year")
}
