% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcInc}
\alias{calcInc}
\title{calcInc}
\usage{
calcInc(rtdat, idat, Args)
}
\arguments{
\item{dat}{Dataset from \code{\link{aggregateInc}}.}

\item{fun}{Function to use for calculations}

\item{calcBy}{Results by Year, Age, or Sex.}
}
\value{
data.frame
}
\description{
Calculates the crude and adjusted incidence.
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
idat <- getBirthDate(Args$inFiles$epifile)
calcInc(rtdat, idat, Args)
}
