% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcInc}
\alias{calcInc}
\title{calcInc}
\usage{
calcInc(rtdat, bdat, Args)
}
\arguments{
\item{rtdat}{Dataset from \code{\link{getRTData}}.}

\item{bdat}{Dataset from \code{\link{getBirthDate}}.}

\item{Args}{provide arguments from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Calculates the crude and adjusted incidence.
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
bdat <- getBirthDate(Args$inFiles$epifile)
calcInc(rtdat, bdat, Args)
}
