% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{calcRubin}
\alias{calcRubin}
\title{calcRubin}
\usage{
calcRubin(est, se)
}
\arguments{
\item{est}{Estimates from m imputations}

\item{se}{Estimates from m imputations}
}
\value{

}
\description{
Calculates standard error according to Rubin's rules.
}
\examples{
e <- c(2, 4) # m = 2
s <- c(0.5, 0.5) # m = 2
x <- calcRubin(e, s)
}
