% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrend.R
\name{calcTrend}
\alias{calcTrend}
\title{calcTrend}
\usage{
calcTrend(dat, wdat = NULL, Formula = "HIVResult ~ Year + AgeCat",
  calcBy = "Year", mergeVars = c("AgeCat"), binom = FALSE,
  fmt = TRUE, ...)
}
\arguments{
\item{dat}{dataset for computing trends.}

\item{wdat}{weights.}

\item{Formula}{string argument for aggregate function; terms and operators must be
separated by white space}

\item{calcBy}{variable(s) to calc the trend by.}

\item{mergeVars}{variables with which to merge \code{dat} and \code{wdat} datasets,
typically \code{AgeCat} variable.}

\item{binom}{use binomial exact formula to calculate CIs: TRUE/FALSE.}

\item{fmt}{convert to percentage and round: TRUE/FALSE.}

\item{stpopVar}{name of standard population column in \code{wdat}}
}
\value{
data.frame
}
\description{
Multipurpose function to calc trends  over time.
}
\examples{
inFiles <- getFiles()
Args <- setArgs(inFiles, 
  AgeCat=c(15, 25, 55), 
  Sex="Fem")
hiv=getHIV(Args)

# Calculate HIV prevalence using KZN 2015 weights
calcTrend(hiv, wdat, Formula="HIVResult ~ Year + AgeCat", mergeVar="AgeCat", calcBy="Year")
}
