% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{combineEst}
\alias{combineEst}
\title{combineEst}
\usage{
combineEst(dat)
}
\arguments{
\item{dat}{takes results from a function (i.e., \code{calcInc}.)}
}
\value{
list
}
\description{
Collect all estimates into single matrix
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
idat <- getBirthDate(Args$inFiles$epifile)
dat <- mclapply(seq(Args$nSimulations), 
  function(i) calcInc(rtdat, idat, Args), 
  mc.cores=Args$mcores)
cdat <- combineEst(dat) 
}
