% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{combineEst}
\alias{combineEst}
\title{combineEst}
\usage{
combineEst(dat, get_names = stdGetNames)
}
\arguments{
\item{dat}{Dataset from \code{\link{setInc}}.}

\item{get_names}{Names of the estimates to be collected.}
}
\value{
list
}
\description{
Collect all estimates into single matrix
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
bdat <- getBirthDate(Args$inFiles$epifile)
calcInc <- setInc(rtdat, bdat, doPoisAge, Args)
dat <- mclapply(seq(Args$nSim), calcInc,
  mc.cores=Args$mcores)
cdat <- combineEst(dat) 
}
