% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{doPoisYear}
\alias{doPoisYear}
\title{doPoisYear}
\usage{
doPoisYear(dat, age_dat = eval.parent(quote(age_dat)))
}
\arguments{
\item{dat}{Dataset from a function \code{\link{getAgeYear}}.}

\item{age_dat}{Dataset from function \code{\link{getAgeYear}}.}
}
\value{
data.frame
}
\description{
Do poisson regression and incidence rates by year.
}
