% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{doPoisYear}
\alias{doPoisYear}
\title{doPoisYear}
\usage{
doPoisYear(dat)
}
\arguments{
\item{dat}{takes dataset from a function (i.e., \code{doIncData}.)}
}
\value{
data.frame
}
\description{
DO poisson regression for incidence rates by year or age.
}
\examples{
doPoisYear <- poisFunc("as.factor(Year)")
doPoisAge <- poisFunc("as.factor(AgeCat)")
}
