% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{dropTasPData}
\alias{dropTasPData}
\title{Drop TasP (Northern Areas) from a dataframe.}
\usage{
dropTasPData(dat, bsdat = NULL)
}
\arguments{
\item{dat}{A dataset, which will be merged with the Bounded Structures
dataset, to determine if observations come from the TasP (northern) areas. 
If an observation cannot be linked to an area, it is kept.}

\item{bsdat}{The Bounded Structures dataset with BSIntID and PIPSA variables.}
}
\value{
data.frame
}
\description{
Function to drop obervations from TasP areas.
}
\examples{
# This is what the dropTasP argument does when it is TRUE.
hiv <- readHIVData(dropTasP = FALSE)
hiv <- dropTasPData(hiv)
}
