% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{dropTasPData}
\alias{dropTasPData}
\title{dropTasPData}
\usage{
dropTasPData(dat, inFile = getFiles()$pipfile)
}
\arguments{
\item{dat}{A dataset.}
}
\value{
data.frame
}
\description{
Function to drop individuals who tested in TasP areas.
}
