% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getAgeYear}
\alias{getAgeYear}
\title{getAgeYear}
\usage{
getAgeYear(dat)
}
\arguments{
\item{dat}{A dataset with an Age and Year variable.}
}
\value{
data.frame
}
\description{
Calculate mean age by year. This is mostly used in \code{MIpredict}
}
\examples{
age_dat <- getAgeYear(dat=setHIV(setArgs()))
}
