% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrend.R
\name{getAggregate}
\alias{getAggregate}
\title{getAggregate}
\usage{
getAggregate(dat, Formula = "IIntID ~ Year")
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{Formula.}{}
}
\value{
list
}
\description{
Gets aggregated numbers by category.
}
\examples{
Args <- setArgs()
hiv <- getHIV()
hiv <- setData(hiv, Args)
getAggregate(hiv, "IIntID ~ Year + HIVResult")
}
\keyword{internal}
