% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTCov.R
\name{getARTDates}
\alias{getARTDates}
\title{getARTDates}
\usage{
getARTDates(inFile = Args$inFiles$epifile)
}
\arguments{
\item{inFile}{path to data, typically \code{Args$inFile}.}
}
\value{
data.frame
}
\description{
get ART initiation dates
}
