% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getARTData.R
\name{getARTDates}
\alias{getARTDates}
\title{getARTDates}
\usage{
getARTDates(dat = getEpisodes())
}
\arguments{
\item{dat}{A dataset which has a variable called \code{EarliestARTInitDate}. If
NULL, the function calls \code{\link{getEpisodes}} with that variable name.}
}
\value{
data.frame
}
\description{
Get ART initiation dates
}
\examples{
getARTDates()
}
