% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{getBirthDate}
\alias{getBirthDate}
\title{getBirthDate}
\usage{
getBirthDate(addVars = " ")
}
\arguments{
\item{addVars}{String for a regular expression to select additional vars}

\item{inFile}{File path to the dataset, default is set to \code{\link{getFiles}}}
}
\value{
data.frame
}
\description{
Gets birth dates from \code{\link{getEpisodes}} data
}
\examples{
Args <- setArgs()
getBirthDate(addVars="Female")
}
