% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{getBirthDate}
\alias{getBirthDate}
\title{getBirthDate}
\usage{
getBirthDate(inFile = Args$inFile$epifile, addVars = " ")
}
\arguments{
\item{inFile}{File path to the dataset, default is set to \code{\link{getFiles}}}

\item{addVars}{string for a regular expression to select additional vars}
}
\value{
data.frame
}
\description{
gets birth dates from \code{\link{getEpisodes}} data
}
