% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{getBSData}
\alias{getBSData}
\title{getBSData}
\usage{
getBSData(inFile = NULL)
}
\arguments{
\item{inFile}{filepath to data.}
}
\value{
data.frame
}
\description{
Read in Bounded Structures data.
}
\examples{
getBSData() 
}
