% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{getBSMax}
\alias{getBSMax}
\title{getBSMax}
\usage{
getBSMax(inFile = getFiles()$epi_rda, minDays = 0)
}
\arguments{
\item{inFile}{file path to Episodes dataset (\code{getFiles()$epi_rda}).}

\item{minDays}{Value of 1:366 min days spent in DSA to qualify as being a resident in
that year.}
}
\value{
data.frame
}
\description{
Gets the BSIntID that IIntID spent most time in a surveillance year.
}
\examples{
getBSMax()
}
