% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getCircum}
\alias{getCircum}
\title{getCircum}
\usage{
getCircum(Keep)
}
\arguments{
\item{Keep}{Keeps or drops circumcised men.}
}
\description{
gets Circumcision data from MGH AHRI datasets.
}
\examples{
keepCircum <- getCircum(Keep=1)
dropCircum <- getCircum(Keep=0)
getCircumcision <- getCircum(Keep = c(0, 1))
}
