% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getCircumcisionData}
\alias{getCircumcisionData}
\title{Gets circumcision data from the MGH dataset.}
\usage{
getCircumcisionData(dat = getMGH())
}
\arguments{
\item{dat}{A dataset from \code{\link{getMGH}}.}
}
\description{
Gets circumcision data from the MGH dataset with the year of
circumcision and an indicator of ever circumcised.
}
\examples{
getCircumcisionData(dat = getMGH())
}
