% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircumcision.R
\name{getCircumcisionData}
\alias{getCircumcisionData}
\title{getCircumcisionData}
\usage{
getCircumcisionData(inFile = getFiles()$mghfile)
}
\arguments{
\item{inFile}{filepath to dataset, default is \code{getFiles()$mghfile}.}
}
\description{
gets Circumcision data from MGH AHRI datasets.
}
