% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getCondomUseData}
\alias{getCondomUseData}
\title{getCondomUseData}
\usage{
getCondomUseData(dat)
}
\arguments{
\item{inFile}{Filepath to dataset, default is \code{getFiles()$mghfile}.}
}
\description{
gets Condom use data from AHRI datasets.
}
\examples{
dat <- readHealthDataMen()
getCondomUseData(dat)
}
