% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVLData.R
\name{getCVLData}
\alias{getCVLData}
\title{getCVLData}
\usage{
getCVLData(Args)
}
\arguments{
\item{Args}{requires Args (\code{\link{setArgs}}) for filepath and setting age}
}
\value{
data.frame
}
\description{
Pulls in Community Viral Load data
}
