% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R, R/getRTData.R
\name{getDates}
\alias{getDates}
\title{getDates}
\usage{
getDates(f)

getDates(f)
}
\arguments{
\item{f}{a function, either \code{min} or \code{max}.}
}
\value{
data.frame

data.frame
}
\description{
Function to get earliest/latest test dates

Function to get earliest/latest test dates
}
