% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTData.R
\name{getDatesMax}
\alias{getDatesMax}
\title{getDatesMax}
\usage{
getDatesMax(dat, Var, Name)
}
\arguments{
\item{dat}{a dataset.}

\item{Var}{a variable name.}

\item{Name}{new variable name.}
}
\value{
data.frame
}
\description{
Function to get latest test dates
}
