% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R, R/getRTData.R
\name{getDatesMin}
\alias{getDatesMin}
\title{getDatesMin}
\usage{
getDatesMin(dat, Var, Name)

getDatesMin(dat, Var, Name)
}
\arguments{
\item{dat}{a dataset.}

\item{Var}{a variable name.}

\item{Name}{new variable name.}
}
\value{
data.frame

data.frame
}
\description{
Function to get earliest test dates

Function to get earliest test dates
}
