% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{getDemResident}
\alias{getDemResident}
\title{getDemResident}
\usage{
getDemResident(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}. Note that the function will drop
residents according to \code{Args$ResRule}.}
}
\value{
data.frame
}
\description{
Gets the total number of residents from the Episodes dataset.
}
