% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{getDemResident}
\alias{getDemResident}
\title{getDemResident}
\usage{
getDemResident(Args, prop = 0.5)
}
\arguments{
\item{Args}{}

\item{prop}{proportion time spent in DSA to be included in analysis.}
}
\value{
data.frame
}
\description{
Gets the total number of residents from the Episodes dataset.
}
