% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{getEpisodes}
\alias{getEpisodes}
\title{getEpisodes}
\usage{
getEpisodes(inFile = getFiles()$epi_rda)
}
\arguments{
\item{inFile}{File path to the dataset, default is set to \code{\link{getFiles}}.}
}
\value{
data.frame
}
\description{
Loads Episodes .Rdata into memory, see \code{\link{readEpisodes}}.
}
