% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{getFiles}
\usage{
getFiles(root = setRoot(),
  hivfile = "Derived/HIVSurveillance/2018/RD05-99_ACDIS_HIV_All.dta",
  hiv_rda = "Derived/Analytics/RD05-99_ACDIS_HIV_All.Rda",
  elifile = "Source/HIVSurveillance",
  eli_rda = "Derived/Analytics/HIVSurveillanceYear.Rda",
  hsefile = "Derived/Analytics/HSE2009_2012.dta",
  wghfile = "Derived/WGH_MGH/2018/RD03-99_ACDIS_WGH_ALL.dta",
  wgh_rda = "Derived/Analytics/RD03-99_ACDIS_WGH_ALL.Rda",
  mghfile = "Derived/WGH_MGH/2018/RD04-99_ACDIS_MGH_ALL.dta",
  mgh_rda = "Derived/Analytics/RD04-99_ACDIS_MGH_ALL.Rda",
  bsifile = "Derived/BoundedStructure/2017/RD01-03_ACDIS_BoundedStructures.dta",
  bsm_rda = "Derived/Analytics/MaxBSIntID.Rda",
  bsc_rda = "Derived/Analytics/BoundedStructures.Rda",
  epi_dta = "Source/Episodes/2018/SurveillanceEpisodesBasicAgeYrHIV.dta",
  epi_rda = "Derived/Analytics/SurveillanceEpisodes.Rda",
  parfile = "Derived/Analytics/PartnerDat.Rdata",
  prvfile = "Derived/Analytics/HIV_Prev_Aug21.csv",
  pipfile = "Derived/Analytics/PIPBoundedStructures2018.dta",
  bscfile = "Source/BoundedStructure/BSIntID_Coords.csv",
  fem_art = "Derived/ARTemis/2017/ART_All.csv",
  mal_art = "Derived/ARTemis/2017/ART_All.csv",
  pvlfile = "Source/CVL_2011_2014/Community Viral Load 2011-2014.dta")
}
\arguments{
\item{root}{file path to derived files}
}
\value{
list
}
\description{
get/set file paths to AHRI data
}
\examples{
inFiles <- getFiles()
}
