% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{getFiles}
\usage{
getFiles(root = file.path(Sys.getenv("HOME"), "Documents/AC_Data"),
  demfile = "Derived/Demography/2016/RD02-01_ACDIS_Demography.csv",
  indfile = "Derived/Individuals/2016/RD01-01_ACDIS_Individuals.csv",
  hiv_dta = "Derived/HIVSurveillance/2017/RD05-99_ACDIS_HIV_All.csv",
  hivfile = "Derived/HIVSurveillance/2017/RD05-99_ACDIS_HIV_All.RData",
  hsefile = "Derived/HSE/HSE2009_2012.dta",
  wghfile = "Derived/WGH_MGH/2016/RD03-99_ACDIS_WGH_ALL.csv",
  mghfile = "Derived/WGH_MGH/2016/RD04-99_ACDIS_MGH_ALL.csv",
  bsifile = "Derived/BoundedStructure/2017/RD01-03_ACDIS_BoundedStructures.dta",
  bsmfile = "Derived/Other/MaxBSIntID.csv",
  epi_dta = "Source/Episodes/2017/SurveillanceEpisodesBasicAgeYrHIV.dta",
  epifile = "Derived/Episodes/2017/SurveillanceEpisodes.Rdata",
  bsafile = "Derived/Other/BS_Area.Rdata",
  parfile = "Derived/Other/PartnerDat.Rdata",
  prvfile = "Derived/Other/HIV_Prev_Aug21.csv",
  pipfile = "Derived/BoundedStructure/PIPBoundedStructures.xlsx",
  bscfile = "Derived/BoundedStructure/BSIntID_Coords.csv",
  pvlfile = "Source/CVL_2011_2014/Community Viral Load 2011-2014.dta")
}
\arguments{
\item{root}{file path to derived files}
}
\value{
list
}
\description{
get/set file paths to AHRI data
}
\examples{
inFiles <- getFiles()
}
