% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{getGImpute}
\alias{getGImpute}
\title{Add theimputed infection times to an existing dataframe.}
\usage{
getGImpute(rtdat, gdat, i)
}
\arguments{
\item{rtdat}{An existing dataset generated from \code{\link{getRTData}}.}

\item{gdat}{The imputed dateset from \code{\link{gImpute}}.}

\item{i}{The ith imputed infection time in gdat.}
}
\value{
data.frame
}
\description{
Adds the ith column of imputed infection times from \code{\link{gImpute}} to an
existing dataset.
}
\examples{
\donttest{
# See full examples at https://github.com/vando026/ahri/wiki/6-G-Imputation
library(IntCens)
data(gdat)
data(rtdat)
home <- file.path('/home/alain/Seafile/AHRI_Data')
write.table(gdat, file=file.path(home, "intcens/input_data.txt"),
  row.names=FALSE, quote=FALSE)
unireg(input = file.path(home, "intcens/input_data.txt"),
  output = file.path(home, "intcens/res_dat.txt"),
  model = "(Time, sero_event) = Age0 + Age2 + EverCircum",
  subject_id = "IIntID", convergence_threshold=0.01, r = 1.0)
ureg_res <- readUniRegResults(
  File=file.path(home, "intcens/res_dat.txt"))
Args <- setArgs(
  Years=c(2005:2018), nSim=50,
  Age=list(Mal=c(15, 54)))
gdates <- gImpute(gdat, ureg_res, nSim=Args$nSim)
gdat_ex <- getGImpute(rtdat, gdates, 1)
gdat_ex[gdat_ex$sero_event==1, ]
}
}
