% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{getHIV}
\alias{getHIV}
\title{getHIV}
\usage{
getHIV(inFile = NULL)
}
\arguments{
\item{inFile}{File path to .rda dataset, default is \code{getFiles()$hiv_rda}. Leave as
NULL if you don't know what to do or see \code{\link{setFiles}}.}
}
\value{
data.frame
}
\description{
Load in the .rda version of the HIV surveillance dataset.
}
