% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{getHIV}
\alias{getHIV}
\title{getHIV}
\usage{
getHIV(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
get all valid test results from HIV surveillance.
}
