% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVCumTest}
\alias{getHIVCumTest}
\title{getHIVCumTest}
\usage{
getHIVCumTest(Args, num_test = 1, dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{num_test}{Integer for number of times tested. Use 2 for incidence cohort.}

\item{dat}{Dataset, otherwise made with \code{\link{setHIVMiss}}.}
}
\value{

}
\description{
Get summary of number of times participants test.
}
