% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVCumTest}
\alias{getHIVCumTest}
\title{getHIVCumTest}
\usage{
getHIVCumTest(dat, ntest = 1)
}
\arguments{
\item{dat}{Dataset from  \code{\link{getHIVEligible}}.}

\item{num_test}{Integer for number of times tested. Use 2 for incidence cohort.}
}
\value{

}
\description{
Get summary of number of times participants test.
}
