% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVEligible}
\alias{getHIVEligible}
\title{getHIVEligible}
\usage{
getHIVEligible(dat = NULL)
}
\arguments{
\item{dat}{Default is Null or loads dataset from \code{\link{setHIVMiss}}.}
}
\value{
data.frame
}
\description{
Get eligibility for HIV testing.
}
