% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVEligible}
\alias{getHIVEligible}
\title{getHIVEligible}
\usage{
getHIVEligible(Args, dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dat}{Dataset, otherwise made with \code{\link{setHIVMiss}}.}
}
\value{

}
\description{
Get eligibility for HIV testing.
}
