% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVIncEligible}
\alias{getHIVIncEligible}
\title{getHIVIncEligible}
\usage{
getHIVIncEligible(Args, ids = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{ids}{keep only IDs.}
}
\value{
data.frame
}
\description{
Get number of participants eligible for HIV incidence cohort. It is
recommended that you use all years in the HIV surveillance system (i.e. from 2003
onward).
}
\examples{
Args <- setArgs(Years=c(2003:2018))
getHIVIncEligible(Args)
}
