% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVIncEligible}
\alias{getHIVIncEligible}
\title{getHIVIncEligible}
\usage{
getHIVIncEligible(Args, dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dat}{Dataset, otherwise made with \code{\link{setHIVMiss}}.}
}
\value{

}
\description{
Get number of participants eligible for HIV incidence cohort.
}
