% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVMiss2}
\alias{getHIVMiss2}
\title{getHIVMiss2}
\usage{
getHIVMiss2(Args, Root = setRoot())
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{Root}{The root path to folder containing HIV surveillance datasets.}
}
\value{

}
\description{
Gets missed test dates but pulls each HIV surveillance dataset by year.
}
