% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{getHIVRefused}
\alias{getHIVRefused}
\title{getHIVRefused}
\usage{
getHIVRefused(dat)
}
\arguments{
\item{dat}{Dataset from \code{\link{getHIVEligible}}.}
}
\value{
data.frame
}
\description{
Get summary of participants that refused to test.
}
