% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncData}
\alias{getIncData}
\title{getIncData}
\usage{
getIncData(rtdat, idat, Args)
}
\arguments{
\item{rtdat}{dataset from \code{\link{getRTData}}.}

\item{idat}{dataset from \code{\link{getBirthDate}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Function that imputes sero date, splits at censoring date and set the age.
}
\examples{
hiv   <- getHIV(Args)
rtdat <- getRTData(hiv)
idat <- getBirthDate(Args$inFiles$epifile)
getIncData(rtdat, idat, Args)
}
