% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncData}
\alias{getIncData}
\title{getIncData}
\usage{
getIncData(rtdat, wdat, Args)
}
\arguments{
\item{rtdat}{dataset from \code{\link{getRTData}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Function used to prepare the data for \code{\link{getIncidence}}.
}
