% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncData}
\alias{getIncData}
\title{getIncData}
\usage{
getIncData(rtdat, bdat, Args, func = identity)
}
\arguments{
\item{rtdat}{dataset from \code{\link{getRTData}}.}

\item{bdat}{dataset from \code{\link{getBirthDate}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}

\item{func}{Function to perform additional operation.}
}
\value{
data.frame
}
\description{
Function that imputes sero date, splits at censoring date and set the age.
}
\examples{
rtdat <- getRTData(dat=getHIV())
getIncData(rtdat, bdat=getBirthDate(), Args=setArgs())
}
