% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncData}
\alias{getIncData}
\title{Function to impute serodate, split at censoring date, and set age.}
\usage{
getIncData(rtdat, bdat = NULL, Args, func = identity)
}
\arguments{
\item{rtdat}{dataset from \code{\link{getRTData}}.}

\item{bdat}{dataset from \code{\link{getBirthDate}}.}

\item{Args}{takes list from \code{\link{setArgs}}.}

\item{func}{Function to perform additional operation.}
}
\value{
data.frame
}
\description{
Function that imputes sero date, splits at censoring date and set the age.
}
\examples{
Args <- setArgs(imputeMethod = imputeRandomPoint)
rtdat <- getRTData(dat=getHIV())
getIncData(rtdat, Args = Args)
}
