% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(Args, formulas = getFormula())
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}.}

\item{formulas}{A list of formulas for the poisson regression models, for the default see
\code{\link{getFormula}}.}
}
\value{

}
\description{
Default function for using \code{mitools} to do incidence rate
estimation.
}
