% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(Args)
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}}

\item{wdat}{data.frame of weights from \code{\link{getWeights}}}

\item{idat}{Individuals dataset of birthdates from \code{\link{getIndDat}}.}
}
\value{
data.frame
}
\description{
Calculates the crude and adjusted incidence.
}
