% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(Args)
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Calculates the crude and adjusted incidence.
}
