% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R, R/intCens.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(
  Args,
  Compute = miCompute(),
  Combine = miCombine(),
  Extract = miExtract()
)

getIncidence(
  Args,
  Compute = miCompute(),
  Combine = miCombine(),
  Extract = miExtract()
)
}
\arguments{
\item{Args}{Takes list from \code{\link{setArgs}}.}

\item{Compute}{Takes list from \code{\link{miCompute}}.}

\item{Combine}{Takes list from \code{\link{miCombine}}.}

\item{Extract}{Takes list from \code{\link{miExtract}}.}
}
\value{
data.frame

data.frame
}
\description{
Calculates the incidence rates.

Calculates the incidence rates.
}
\keyword{internal}
