% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{getIncidence}
\usage{
getIncidence(Args = setArgs(), sformula = "", aggFun = NULL, newdata = NULL)
}
\arguments{
\item{Args}{takes list from \code{\link{setArgs}}.}

\item{sformula}{A list of formulas for the poisson regression models.}

\item{aggFun}{Function to aggregate sero events and person-time by, see
\code{link{AggFunc}}.}

\item{newdata}{Dataset of variables to predict incidence, see \code{MIpredict}.}
}
\value{
list
}
\description{
Calculate annual HIV incidence rates using the single random-point
approach and multiple imputation.
}
