% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndividualData.R
\name{getIndDat}
\alias{getIndDat}
\title{getIndDat}
\usage{
getIndDat(inFile = getFiles()$indfile)
}
\arguments{
\item{infile}{requires Args, see \code{\link{setArgs}}}
}
\value{
data.frame
}
\description{
Pull in Individuals dataset
}
