% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{getMGH}
\alias{getMGH}
\title{getMGH}
\usage{
getMGH(inFile = NULL)

getMGH(inFile = NULL)
}
\arguments{
\item{inFile}{Filepath to .rda dataset, default is \code{getFiles()$mgh_rda}. Leave as
NULL if you don't know what to do or see \code{\link{setFiles}}.}
}
\value{
data.frame

data.frame
}
\description{
Reads in Men's general health data.

Reads in Men's general health data.
}
