% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMortality.R
\name{getMortalityData}
\alias{getMortalityData}
\title{getMortalityData}
\usage{
getMortalityData(Args, startVar = "HIVPositive", dropHIVPos = FALSE)
}
\arguments{
\item{Args}{see \code{\link{setArgs}}.}

\item{startDate}{string variable at which person-time starts. Use 
HIVPositive for AIDS-related mortality, otherwise only those with an HIVNegative test
or those with the EarliestTest from HIVSurveillance. Or use ObservationStart from the Episodes dataset.}
}
\value{
data.frame
}
\description{
gets mortality data.
}
