% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTData.R
\name{getRTData}
\alias{getRTData}
\title{getRTData}
\usage{
getRTData(dat, onlyRT = TRUE)
}
\arguments{
\item{dat}{dataset from \code{\link{getHIV}}.}

\item{onlyRT}{Drops IDs who are not repeat-testers.}
}
\value{
data.frame
}
\description{
Get all repeat testers from HIV surveillance.
}
\examples{
hiv <- getHIV(Args)
rtdat <- getRTData(hiv)
}
