% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R, R/getRTData.R
\name{getRTData}
\alias{getRTData}
\title{getRTData}
\usage{
getRTData(dat = NULL, onlyRT = TRUE)

getRTData(dat = NULL, onlyRT = TRUE)
}
\arguments{
\item{dat}{dataset from \code{\link{getHIV}}.}

\item{onlyRT}{Drops IDs who are not repeat-testers.}
}
\value{
data.frame

data.frame
}
\description{
Get all repeat testers from HIV surveillance.

Get all repeat testers from HIV surveillance.
}
\examples{
rtdat <- getRTData(dat=getHIV())
rtdat <- getRTData(dat=getHIV())
}
