% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{getYearDates}
\alias{getYearDates}
\title{getYearDates}
\usage{
getYearDates(Years)
}
\arguments{
\item{Years}{The years by which to split the data, for example, `c(2005:2019)`.}
}
\value{
vector
}
\description{
gets numeric dates for 01JanXXXX.
}
