% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R, R/setData.R
\name{makeAgeVars}
\alias{makeAgeVars}
\title{makeAgeVars}
\usage{
makeAgeVars(dat)

makeAgeVars(dat)
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{age_cut}{Vector of ages to make age categories.}

\item{dat}{dataset for which age is needed at a given episode.}
}
\value{
data.frame

data.frame
}
\description{
Centers age variable and takes square and cube.

Centers age variable and takes square and cube.
}
\examples{
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
}
