% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{makeAgeVars}
\alias{makeAgeVars}
\title{makeAgeVars}
\usage{
makeAgeVars(dat, time2 = NULL, age_cut = NULL, birthdate = NULL)
}
\arguments{
\item{dat}{A dataset}

\item{time2}{The name of a date variable that is used to calculate the Age variable using the
\code{\link{getBirthDate}} function. Age is calculated as (time2 - birthdate)/365.35.}

\item{age_cut}{Vector of ages to make age categories from \code{link{setArgs}}.}

\item{birthdate}{Dataset of birthdates, if NULL it uses \code{\link{getBirthDate}}.}
}
\value{
data.frame
}
\description{
Function for making Age and AgeCat variables
}
\examples{
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
}
