% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeData.R
\name{makeAgeVars}
\alias{makeAgeVars}
\title{makeAgeVars}
\usage{
makeAgeVars(dat)
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}
}
\value{
data.frame
}
\description{
Centers age variable and takes square and cube.
}
\examples{
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
}
