% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHealthData.R
\name{makeAgeVars}
\alias{makeAgeVars}
\title{makeAgeVars}
\usage{
makeAgeVars(dat, age_cut = c(15, 25, 35, 100))
}
\arguments{
\item{dat}{dataset for which age is needed at a given episode.}

\item{age_cut}{Vector of ages to make age categories.}
}
\value{
data.frame
}
\description{
Centers age variable and takes square and cube.
}
\examples{
adat <- setAge(sdata)
adat <- makeAgeVars(adat)
}
