% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{MIdata}
\alias{MIdata}
\title{MIdata}
\usage{
MIdata(rtdat, Args, f = identity, bdat = NULL)
}
\arguments{
\item{rtdat}{Dataframe from \code{\link{getRTData}}.}

\item{Args}{Takes list from \code{\link{setArgs}}.}

\item{f}{Function to perform additional operations.}

\item{bdat}{Dataframe of birthdates, otherwise it use \code{\link{getBirthDate}}}
}
\value{
list
}
\description{
Function to generate dataset with imputed serodate.
}
\examples{
Args <- setArgs(nSim=2, mcores=1)
rtdat <- getRTData(dat=getHIV())
MIdata(rtdat, Args)
}
