% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSData.R
\name{mkBSData}
\alias{mkBSData}
\title{mkBSData}
\usage{
mkBSData(inFiles)
}
\arguments{
\item{inFile}{file path to import Demography dataset.}

\item{outFile}{file path to write dataset.}
}
\value{
data.frame
}
\description{
Makes a specific file for the IntCens project
}
