% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{mkHIVTestTable}
\alias{mkHIVTestTable}
\title{mkHIVTestTable}
\usage{
mkHIVTestTable(Args)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Make a table of HIV test participation.
}
