% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInc.R
\name{plotIncPrev}
\alias{plotIncPrev}
\title{plotIncPrev}
\usage{
plotIncPrev(
  inc,
  prev = NULL,
  x = NULL,
  yLim = 8,
  Main = "",
  prev_lty = 1,
  Args = NULL,
  gfun = png,
  bottom = TRUE,
  left = TRUE,
  right = TRUE,
  bwidth = 2.5,
  fname = "test",
  inc_col = "blue",
  prev_col = "red"
)
}
\arguments{
\item{inc}{Incidence data, must include CIs, from \code{\link{getIncidence}}.}

\item{prev}{Prevalence data, only a vector.}
}
\description{
Plots incidence and prevalence.
}
\examples{
\donttest{
inc <- getIncidence(Args)$adj$Year
hiv <- setHIV(Args) 
prev <-  calcTrend(hiv, 
  Formula = "HIVResult ~ Year", fmt=FALSE)
prev <- prev$adj.rate*100
plotIncPrev(inc, prev) 
}
}
\keyword{internal}
