% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{readEpisodes}
\alias{readEpisodes}
\title{readEpisodes}
\usage{
readEpisodes(
  inFile = getFiles()$epi_dta,
  outFile = getFiles()$epi_rda,
  dropTasP = TRUE,
  Vars = " ",
  write_rda = TRUE
)
}
\arguments{
\item{inFile}{File path to the .dta dataset, default is set to \code{\link{getFiles}}.}

\item{outFile}{File path to the write the .Rda dataset, default is set to \code{\link{getFiles}}.}

\item{dropTasP}{default is to drop TasP areas.}

\item{Vars}{A regular expression string.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads in the Surveillance Episodes dataset
}
\examples{
readEpisodes(Vars="ART", dropTasP=TRUE)
}
