% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{readEpisodes}
\alias{readEpisodes}
\title{readEpisodes}
\usage{
readEpisodes(inFile = Args$inFiles$epi_dta,
  outFile = Args$inFiles$epifile)
}
\arguments{
\item{inFile}{File path to the dataset, default is set to \code{\link{getFiles}}.}
}
\value{
data.frame
}
\description{
Reads in new Episodes dta dataset which replaces the Demography dataset (for
2017) and converts it to a .Rdata file.
}
\examples{
Args <- setArgs()
readEpisodes()
}
