% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{readEpisodes}
\alias{readEpisodes}
\title{readEpisodes}
\usage{
readEpisodes(
  inFile = NULL,
  outFile = NULL,
  dropTasP = TRUE,
  addVars = " ",
  write_rda = TRUE,
  nrow = Inf
)
}
\arguments{
\item{inFile}{File path to the .dta dataset, default is set to \code{\link{setFiles}}.}

\item{outFile}{File path to the write the .Rda dataset, default is set to \code{\link{setFiles}}.}

\item{dropTasP}{default is to drop TasP areas.}

\item{addVars}{A regular expression string representing the variables to be added.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads in the Surveillance Episodes dataset
}
\examples{
\donttest{
readEpisodes(addVars="CurrentlyEmployed|UnEmployment")
readEpisodes(dropTasP=FALSE, addVars="Employ")
}
}
