% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{readEpisodes}
\alias{readEpisodes}
\title{readEpisodes}
\usage{
readEpisodes(inFile = NULL, dropTasP = TRUE, addVars = " ", write_rda = TRUE)
}
\arguments{
\item{inFile}{File path to the .dta dataset, default is set to \code{\link{setFiles}}.}

\item{dropTasP}{default is to drop TasP areas.}

\item{addVars}{A regular expression string representing the variables to be added.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads in the Surveillance Episodes dataset. Note, this function drops
participants with missing values for \code{Sex} (there are a handful of these).
}
\examples{
\donttest{
readEpisodes(addVars="CurrentlyEmployed|UnEmployment")
readEpisodes(dropTasP=FALSE, addVars="Employ")
}
}
