% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(
  inFile = NULL,
  dropTasP = TRUE,
  addVars = " ",
  drop15Less = TRUE,
  write_rda = TRUE
)
}
\arguments{
\item{inFile}{File path to .dta, default is set by \code{\link{setFiles}}.}

\item{dropTasP}{Default is to drop TasP surveillance areas from the data.}

\item{addVars}{A regular expression string representing the variables to be added.}

\item{drop15Less}{Default is to drop all observations with Age < 15 years. Only
participants >=15 years eligible for HIV testing.}

\item{write_rda}{Default is to write the .Rda file.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
\examples{
# Writes .Rda to file
\donttest{
readHIVData(write_rda = TRUE)
}
# Drop TasP areas, dont write to file
readHIVData(dropTasP = FALSE, write_rda = TRUE)
# Saves to global environment 
hdat <- readHIVData(write_rda=FALSE)
# Add variables
hdat <- readHIVData(addVars="HIVRefused|WhereLastTested", write_rda=FALSE)
}
