% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.r
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(
  inFile = getFiles()$hivfile,
  outFile = getFiles()$hiv_rda,
  dropTasP = TRUE
)
}
\arguments{
\item{dropTasP}{Drop TasP surveillance areas from the data.}

\item{inFiles}{File path from \code{\link{getFiles}}.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
