% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(inFile = getFiles()$hivfile, dropTasP = TRUE)
}
\arguments{
\item{dropTasP}{drop TasP surveillance areas from the data.}

\item{inFiles}{file paths from \code{\link{setArgs}}.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
