% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(inFile = getFiles()$hivfile, outFile = getFiles()$hiv_rda,
  dropTasP = TRUE)
}
\arguments{
\item{inFile}{File path to .dta, see \code{\link{setFiles}}.}

\item{outFile}{File path to save .Rda, see \code{\link{setFiles}}.}

\item{dropTasP}{Drop TasP surveillance areas from the data.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
