% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(
  inFile = getFiles()$hivfile,
  outFile = getFiles()$hiv_rda,
  dropTasP = TRUE,
  write_rda = TRUE
)
}
\arguments{
\item{inFile}{File path to .dta, default is set by \code{\link{setFiles}}.}

\item{outFile}{File path to save .Rda, default is set by \code{\link{setFiles}}.}

\item{dropTasP}{Drop TasP surveillance areas from the data.}

\item{write_rda}{Default is to write the .Rda file.}

\item{addVars}{A regular expression string representing the variables to be added.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
