% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{readHIVData}
\alias{readHIVData}
\title{readHIVData}
\usage{
readHIVData(inFiles = getFiles(), dropTasP = TRUE)
}
\arguments{
\item{inFiles}{file paths from \code{\link{setArgs}}.}

\item{dropTasP}{drop TasP surveillance areas from the data.}
}
\value{
data.frame
}
\description{
Reads in AHRI data from csv file
}
