% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{readUniRegResults}
\alias{readUniRegResults}
\title{Read the results from the unireg model into dataframe.}
\usage{
readUniRegResults(File = NULL)
}
\arguments{
\item{File}{File path to \code{\link{uniReg}} output.txt.}
}
\description{
Read the results from the \code{\link{uniReg}} model into .Rdata
format.
}
\examples{
# install.packages(system.file("IntCens_0.2.tar.gz", package = "ahri"), repos = NULL, type = 'source')
library(IntCens)
data(gdat_demo)
home <- file.path('/home/alain/Seafile/AHRI_Data')
write.table(gdat_demo, file=file.path(home, "intcens/input_data.txt"),
  row.names=FALSE, quote=FALSE)
unireg(input = file.path(home, "intcens/input_data.txt"),
  output = file.path(home, "intcens/res_dat.txt"),
  model = "(Time, sero_event) = Age0 + Age2 + EverCircum",
  subject_id = "IIntID", convergence_threshold=0.01, r = 1.0)
ureg_res <- readUniRegResults(
  File=file.path(home, "intcens/res_dat.txt"))
}
