% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{runUniReg}
\alias{runUniReg}
\title{runUniReg}
\usage{
runUniReg(Vars, Name)
}
\arguments{
\item{Vars}{Variables to feed into uniReg model.}

\item{Name}{Name of input txt file.}
}
\value{
list
}
\description{
Helper function to run \code{\link{uniReg}}.
}
