% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setArgs.R
\name{setArgs}
\alias{setArgs}
\title{setArgs}
\usage{
setArgs(inFiles = getFiles(), Years = c(2004:2017), Age = list(All =
  c(15, 54)), AgeCat = NULL, AgeBy = 5, ResRule = 0, nSim = 500,
  imputeMethod = imputeRandomPoint, aname = "filename",
  printout = FALSE, mcores = 1, MoreArgs = NULL)
}
\arguments{
\item{inFiles}{list of files to pass to function.}

\item{Years}{numeric vector of years.}

\item{Age}{list of ages as in \code{list(Fem=c(15, 49), Mal=c(15, 54)}.}

\item{AgeCat}{numeric vector of age categories.}

\item{ResRule}{proportion of time spent in DSA, greater than >.}

\item{nSim}{number of simulations to perform.}

\item{imputeMethod}{select either \code{\link{imputeRandomPoint}} or
\code{\link{imputeMidPoint}}.}

\item{aname}{root name to associate with output or filenames.}

\item{printout}{print out results.}

\item{mcores}{number of cores to use for parallel package (used in
\code{\link{getIncidence}}).}

\item{MoreArgs}{a list, which adds more arguments if needed.}
}
\value{
list
}
\description{
Set the arguments for functions.
}
\examples{
inFiles <- getFiles()
Args <- setArgs(inFiles, Years=c(2004:2015), All=list(Mal=c(15, 54)))
Args <- setArgs(inFiles, Years=c(2004:2015), MoreArgs=list(knots=2))
}
