% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{setData}
\alias{setData}
\title{setData}
\usage{
setData(dat, Args, time2 = NULL, birthdate = NULL)
}
\arguments{
\item{dat}{Dataset for which age is needed at a given episode.}

\item{Args}{takes a list from \code{\link{setArgs}}.}

\item{time2}{A date variable that is used to calculate age from Birthdate to time2.}
}
\value{
data.frame
}
\description{
Sets data according to values in \code{\link{setArgs}}.
}
\examples{
rtdata <- getRTData(hiv)
sdat <- splitAtEarlyPos(rtdat)
adat <- setData(sdat, Args)
}
