% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{setEpisodes}
\alias{setEpisodes}
\title{setEpisodes}
\usage{
setEpisodes(Args, dropTasP = TRUE)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dropTasP}{default is to drop TasP areas.}
}
\value{
data.frame
}
\description{
set episodes data according to Args and drops TasP Areas if needed.
}
\examples{
setEpisodes(setArgs())
}
