% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpisodes.R
\name{setEpisodes}
\alias{setEpisodes}
\title{setEpisodes}
\usage{
setEpisodes(Args = setArgs(), dat = NULL)
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}

\item{dat}{A dataset generated from \code{\link{readEpisodes}}, which exists in the
global environment. If NULL, it reads in the corresponding .Rda file (see
\code{\link{setFiles}}.}
}
\value{
data.frame
}
\description{
Set the Episodes data according to Arguments.
}
\examples{
Args <- setArgs(Years=c(2005:2010))
setEpisodes(Args)
epidat <- readEpisodes(write_rda=FALSE)
setEpisodes(Args, epidat)
}
