% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIV.R
\name{setHIV}
\alias{setHIV}
\title{setHIV}
\usage{
setHIV(Args = setArgs())
}
\arguments{
\item{Args}{requires Args, see \code{\link{setArgs}}.}
}
\value{

}
\description{
set HIV data according arguments of \code{\link{setArgs}}.
}
