% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHIVMiss.R
\name{setHIVMiss}
\alias{setHIVMiss}
\title{setHIVMiss}
\usage{
setHIVMiss(
  inFile = getFiles()$elifile,
  outFile = getFiles()$eli_rda,
  dropTasP = TRUE
)
}
\arguments{
\item{inFile}{The root path to folder containing HIV surveillance datasets.}

\item{outFile}{Path to write new dataset.}

\item{dropTasP}{Drop TasP surveillance areas from the data.}
}
\value{
data.frame
}
\description{
Gets HIV test dates by merging yearly HIV surveillance datasets.
}
\keyword{internal}
