% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitAtEarlyPos}
\alias{splitAtEarlyPos}
\title{splitAtEarlyPos}
\usage{
splitAtEarlyPos(dat = NULL, splitYears = NULL)
}
\arguments{
\item{dat}{a dataset}

\item{splitYears}{Vector from \code{\link{getYearDates}}.}
}
\value{
data.frame
}
\description{
Split data at the earliest HIV-positive date.
}
\examples{
hiv <- getHIV(Args)
rtdat <- getRTData(hiv)
sdat <- imputeMidPoint(rtdat)
splitAtEarlyPos(sdat, splitYears=Args$Years)
}
