% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitAtSeroDate}
\alias{splitAtSeroDate}
\title{splitAtSeroDate}
\usage{
splitAtSeroDate(dat = NULL, splitYears = c(2003:2018))
}
\arguments{
\item{dat}{a dataset}

\item{splitYears}{Vector from \code{\link{getYearDates}}.}
}
\value{
data.frame
}
\description{
Split data at the imputed seroconversion date.
}
\examples{
hiv <- getHIV(Args)
rtdat <- getRTData(hiv)
sdat <- imputeMidPoint(rtdat)
splitAtSeroDate(sdat, splitYears=Args$Years)
}
