% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{splitData}
\usage{
splitData(dat = NULL, splitYears = NULL, svar = "sero_date",
  len_drop = NULL)
}
\arguments{
\item{dat}{dataset from imputation method, eg \code{\link{imputeRandomPoint()}}.}

\item{svar}{varname to create obs_end variable. Can either split at early_pos or at
sero_date}

\item{years}{to split}
}
\value{
data.frame
}
\description{
Split the data into year episodes at svar
}
\examples{
rtdat <- getRTData(hiv)
sdat <- imputeMidPoint(rtdat)
sdat <- rename(sdat, sero_date=s1)
splitData(sdat, splitYears=Args$Years)
}
