% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData2}
\alias{splitData2}
\title{splitData2}
\usage{
splitData2(dat, years = c(2003:2018))
}
\arguments{
\item{dat}{Dataset must have variables called obs_start, obs_end and event.}
}
\value{
data.frame
}
\description{
version 2 of \code{\link{splitData}}.
}
