% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData2}
\alias{splitData2}
\title{Function for splitting data into episodes.}
\usage{
splitData2(dat, years = NULL)
}
\arguments{
\item{dat}{Dataset must have variables called obs_start, obs_end and event.}

\item{years}{The years by which to split the data. Default is c(2003:2020).}
}
\value{
data.frame
}
\description{
Function for splitting data into episodes.
}
