% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{uniReg}
\alias{uniReg}
\title{uniReg}
\usage{
uniReg(
  xpath,
  InFile,
  OutFile,
  Model,
  ID = NULL,
  inf = "Inf",
  iter = 5000,
  cthresh = 1e-04,
  r = 1,
  printout = FALSE,
  ign_stout = TRUE
)
}
\arguments{
\item{xpath}{The path to the unireg executable.}

\item{InFile}{txt file to be input}

\item{OutFile}{txt file to be output}

\item{Model}{equation to be given}

\item{ID}{name of subject ID}

\item{inf}{Value for infinite, default is "Inf"}

\item{iter}{Number of iterations}

\item{cthresh}{Threshold for convergence}

\item{r}{Threshold for convergence}

\item{printout}{Print results to screen}

\item{ign_stout}{For Linux systems}
}
\description{
Wrapper for Intcens executable by Zeng et al 2016. See
http://dlin.web.unc.edu/software/intcens/ to download the intcens program
for R.
}
\keyword{internal}
