% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{UniReg}
\alias{UniReg}
\title{UniReg}
\usage{
UniReg(InFile, OutFile, Model, ID, iter = 1000, cthresh = 0.001, r = 0)
}
\arguments{
\item{InFile}{txt file to be input}

\item{OutFile}{txt file to be output}

\item{Model}{equation to be given}

\item{ID}{name of subject ID}
}
\description{
wrapper for IntCens fuction
}
\examples{
UniReg(
InFile = file.path(output, "TestSim.txt"),
OutFile = file.path(output, "TestSim_Out.txt"), r = 0.0,
Model = "(time, d) = v1 + v2 + v3",
ID = "id")
}
