% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCens.R
\name{uniRegOne}
\alias{uniRegOne}
\title{uniRegOne}
\usage{
uniRegOne(Vars, Name)
}
\arguments{
\item{Vars}{Vector of RHS character varnames.}

\item{Name}{File name of the gModel results from \code{\link{readUniRegResults}}. Download and install the IntCens library for R from \link{https://dlin.web.unc.edu/software/intcens/}.}
}
\description{
Run gModel on each variable and make table
}
\examples{
\donttest{uniRegOne(c("Age0", "Age2"), Name="icens_mal")}
}
\keyword{internal}
